jobs/webcompat-kb/requirements.txt
click==8.0.4
flake8==3.8.4
google-cloud-bigquery==3.11.4
pytest==6.0.2
pytest-black==0.3.11
pytest-flake8==1.0.6
